+++ rayTracer/src/emscriptenRunner.h
@@ -17,8 +17,7 @@
 #include "renderer.h"
 #include "world.h"
 
-class EmscriptenRunner
-{
+class EmscriptenRunner {
 private:
   static const int PIXELS_PER_BATCH = 40000;
   //   unsigned int startCurrentBatch = 0;
@@ -28,15 +27,15 @@ private:
   std::vector<std::pair<int, int>> pixelsToRender;
 
   int sqrtRaysPerPixel;
-  float halfSubPixelSize;
-  const float STEP_SIZE = 0.05f;
+  double halfSubPixelSize;
+  const double STEP_SIZE = 0.05f;
 
   std::random_device rd;
   std::mt19937 g;
 
   // void render_(EmscriptenRunner *runnerPtr);
 
-  void moveCamera(float posChange);
+  void moveCamera(double posChange);
 
 public:
   EmscriptenRunner(/* args */);
